Engine & Irrlicht: Add double property to save/load data

+++ Projects/Skylicht/Engine/Serializable/CValueProperty.h
@@ -37,7 +37,7 @@ namespace Skylicht
 		Integer,
 		UInteger,
 		Float,
-		Decimal,
+		Double,
 		DateTime,
 		Bool,
 		Vector3,
